automation2/requirements.txt
allure-pytest==2.9.44
allure-python-commons==2.9.44
attrs==21.2.0
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-rerunfailures==10.2
PyYAML==5.4.1
selenium==3.141.0
six==1.16.0
toml==0.10.2
urllib3==1.26.7
